requirements.txt
appdirs==1.4.4
appnope==0.1.2
argon2-cffi==20.1.0
astroid==2.4.2
async-generator==1.10
attrs==20.3.0
authentication==1.1.0
backcall==0.2.0
beautifulsoup4==4.9.3
bleach==3.2.1
boto3==1.9.159
botocore==1.12.159
cached-property==1.5.2
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
Click==7.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.2.1
decorator==4.4.2
defusedxml==0.6.0
distlib==0.3.1
Django==2.2.11
django-admin-tools==0.8.1
django-choices==1.7.1
django-cloudmailin==0.1.1
django-constance==2.8.0
django-controlcenter==0.2.9
django-cors-headers==3.0.2
django-crispy-forms==1.7.2
django-excel==0.0.10
django-extensions==2.1.7
django-filter==2.1.0
django-picklefield==3.0.1
django-pkgconf==0.3.0
django-rest-framework==0.1.0
django-rest-passwordreset==1.1.0
django-rest-swagger==2.2.0
django-restframework==0.0.1
django-storages==1.9.1
django-tables2==2.0.6
django-tenant-schemas==1.10.0
djangorestframework==3.9.4
djangorestframework-api-key==2.0.0
djangorestframework-simplejwt==4.6.0
docutils==0.14
entrypoints==0.3
et-xmlfile==1.0.1
filelock==3.0.12
flake8==3.8.3
Flask==1.1.2
google==3.0.0
google-api-core==1.22.2
google-api-python-client==1.12.1
google-auth==1.21.2
google-auth-httplib2==0.0.4
googleapis-common-protos==1.52.0
gunicorn==20.0.4
httplib2==0.19.0
idna==2.10
ipykernel==5.4.2
ipython==7.5.0
ipython-genutils==0.2.0
ipywidgets==7.6.2
isodate==0.6.0
isort==5.6.4
itsdangerous==1.1.0
itypes==1.2.0
jdcal==1.4.1
jedi==0.18.0
Jinja2==2.11.2
jmespath==0.10.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.7
jupyter-console==6.2.0
jupyter-core==4.7.0
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
jwt==1.1.0
lambda-packages==0.20.0
lazy-object-proxy==1.4.3
lml==0.1.0
lxml==4.5.2
MarkupSafe==1.1.1
mccabe==0.6.1
mistune==0.8.4
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
ndg-httpsclient==0.5.1
nest-asyncio==1.4.3
notebook==6.1.6
oauthlib==3.1.0
openapi-codec==1.3.2
openpyxl==3.0.3
ordered-set==4.0.2
packaging==20.8
pandocfilters==1.4.3
parso==0.8.1
passlib==1.7.4
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.0.0
pipenv==2020.11.15
progressbar==2.5
prometheus-client==0.9.0
prompt-toolkit==2.0.9
protobuf==3.15.6
psycopg2-binary==2.8.6
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
pycryptodome==3.9.9
pyexcel==0.6.6
pyexcel-io==0.6.4
pyexcel-webio==0.1.4
pyflakes==2.2.0
Pygments==2.7.3
PyJWT==2.0.0
pylint==2.6.0
pyOpenSSL==20.0.1
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.6.1
python-docx==0.8.10
python-slugify==1.2.4
python-social-auth==0.3.6
python3-openid==3.2.0
pytz==2019.1
PyYAML==5.1
pyzmq==20.0.0
qtconsole==5.0.1
QtPy==1.9.0
receita-tools==2.2.0
redis==3.5.3
requests==2.24.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rsa==4.7.2
s3transfer==0.2.1
Send2Trash==1.5.0
sentry-sdk==0.17.5
simplejson==3.17.2
six==1.15.0
soupsieve==2.2.1
sqlparse==0.4.1
style==1.1.0
terminado==0.9.1
testpath==0.4.4
texttable==1.6.3
toml==0.10.2
tornado==6.1
traitlets==5.0.5
unicodecsv==0.14.1
Unidecode==1.1.2
update==0.0.1
uritemplate==3.0.1
urllib3==1.25.11
virtualenv==20.2.1
virtualenv-clone==0.5.4
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
WTForms==2.3.3
zappa-django-utils==0.4.0
zeep==3.4.0
